Data_False_Positives/requirements.txt
confuse==1.3.0
pandas==1.1.3
pysmb==1.2.4
python-dotenv==0.14.0
dask[complete]

